Refactor: move PPMAPI data access layer to its own project
