+++ PPM.sln
@@ -37,6 +37,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AuthServiceDataAccess", "Au
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AuthServiceServiceLayer", "AuthServiceServiceLayer\AuthServiceServiceLayer.csproj", "{7090C00E-436C-41EB-80A1-97039DEA335A}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PPMAPIDataAccess", "PPMAPIDataAccess\PPMAPIDataAccess.csproj", "{9C3D80BB-BA4B-4426-8CFC-BB3C29ADB267}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -83,6 +85,10 @@ Global
 		{7090C00E-436C-41EB-80A1-97039DEA335A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7090C00E-436C-41EB-80A1-97039DEA335A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7090C00E-436C-41EB-80A1-97039DEA335A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9C3D80BB-BA4B-4426-8CFC-BB3C29ADB267}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9C3D80BB-BA4B-4426-8CFC-BB3C29ADB267}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9C3D80BB-BA4B-4426-8CFC-BB3C29ADB267}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9C3D80BB-BA4B-4426-8CFC-BB3C29ADB267}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
