--- PPMAPI/DataAccess/DbTableQueries/ValueDecreasesQueries/IValueDecreasesQueries.cs
@@ -1,13 +1,0 @@
-﻿using PPMModelLibrary.Models.ValueModifiers;
-
-namespace PPMAPI.DataAccess.DbTableQueries.ValueDecreasesQueries
-{
-    public interface IValueDecreasesQueries
-    {
-        public void AddValueDecrease(ValueDecrease valueDecrease);
-        public ValueDecrease GetValueDecreaseById(int id);
-        public List<ValueDecrease> GetValueDecreaseByPropertyId(Guid id);
-        public void UpdateValueDecrease(ValueDecrease valueDecrease);
-        public void DeleteValueDecrease(int id);
-    }
-}
