--- PPMAPI/DataAccess/DbTableQueries/ValueIncreasesQueries/IValueIncreasesQueries.cs
@@ -1,13 +1,0 @@
-﻿using PPMModelLibrary.Models.ValueModifiers;
-
-namespace PPMAPI.DataAccess.DbTableQueries.ValueIncreasesQueries
-{
-    public interface IValueIncreasesQueries
-    {
-        public void AddValueIncrease(ValueIncrease valueIncrease);
-        public ValueIncrease GetValueIncreaseById(int id);
-        public List<ValueIncrease> GetValueIncreaseByPropertyId(Guid id);
-        public void UpdateValueIncrease(ValueIncrease valueIncrease);
-        public void DeleteValueIncrease(int id);
-    }
-}
