+++ PPMAPIDataAccess/DbTableQueries/AddressQueries/IAddressesQueries.cs
@@ -1,0 +1,12 @@
+﻿using PPMModelLibrary.Models.UtilityModels;
+
+namespace PPMAPIDataAccess.DbTableQueries.AddressQueries
+{
+    public interface IAddressesQueries
+    {
+        public void AddAddress(Address address);
+        public Address GetAddressById(int id);
+        public void UpdateAddress(Address address);
+        public void DeleteAddress(int id);
+    }
+}
