+++ PPMAPIDataAccess/DbTableQueries/RentablePropertiesQueries/IRentablePropertiesQueries.cs
@@ -1,0 +1,13 @@
+﻿using PPMModelLibrary.Models.Properties;
+
+namespace PPMAPIDataAccess.DbTableQueries.RentalPropertiesQueries
+{
+    public interface IRentalPropertiesQueries
+    {
+        public void AddRentalProperty(RentalProperty rentalProperty);
+        public void DeleteRentalProperty(Guid id);
+        public void UpdateRentalProperty(RentalProperty rentalProperty);
+        public RentalProperty GetRentalPropertyById(Guid id);
+        public List<RentalProperty> GetRentalPropertiesByOwnerId(string id);
+    }
+}
