+++ PPMAPIDataAccess/appsettings.json
@@ -1,0 +1,12 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "APIConnString": "Server=(localdb)\\MSSQLLocalDB;Database=PPMAPI;Trusted_Connection=True;"
+  }
+}
