+++ PPMAPITests/PropertiesQueriesTests.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
-using PPMAPI.DataAccess;
-using PPMAPI.DataAccess.DbTableQueries.PropertiesQueries;
+using PPMAPIDataAccess;
+using PPMAPIDataAccess.DbTableQueries.PropertiesQueries;
 using PPMModelLibrary.Models.UtilityModels;
 using PPMModelLibrary.Models.Properties;
 using PPMModelLibrary.Models.Transactions;
